Move the Righter interfaces to rights, where they belong
